backend/requirements.txt
flask
flask-cors
numpy
pandas
flask-sqlalchemy
psycopg2-binary
bcrypt
flask-mail
pyjwt
flask-migrate
dotenv